Rename mapping-tree-view to cycle-task-objects
