+++ src/ggrc-client/js/components/workflow/cycle-task-objects/cycle-task-objects.stache
